% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate grain-size mean (M1)}
\usage{
calc_mean(data, classes, unit = "phi", method = "arithmetic")
}
\arguments{
\item{data}{\code{Numeric} vector or matrix, grain-size 
distribution(s) to be evaluated.}

\item{classes}{\code{Numeric} vector, grain-size class limits 
corresponding to the input data set.}

\item{unit}{\code{Character} value, unit of the grainsize classes. One 
out of \code{"phi"} and \code{"mu"}, default is \code{"phi"}.}

\item{method}{\code{Character} value, method used for caluclation. One 
out of \code{"arithmetic"} and \code{"geometric"}, default is 
\code{"arithmetric"}.}
}
\value{
\code{Numeric} vector, mean grain-size.
}
\description{
The function calculates the first moment of grain-size distributions, 
using either the arithmetic or geometric method.
}
\examples{

## load example data set
data(LA950)

## calculate arithemtric mean in metric scale
calc_mean(data = LA950$y, 
          classes = LA950$x, 
          unit = "mu")
          
## calculate arithmetic mean in phi-scale
calc_mean(data = LA950$y, 
          classes = convert_units(mu = LA950$x), 
                                  unit = "phi")
                                  
## calculate geometric mean
calc_mean(data = LA950$y, 
          classes = LA950$x, 
          unit = "mu", 
          method = "geometric")
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
