% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize-package.R
\docType{data}
\name{CamSizerXT}
\alias{CamSizerXT}
\title{CamSizerXT example data set}
\format{List of 4
         $ x   : num [1:256] 1 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8 1.9 ...
         $ y   : num [1:256] 0.002 0 0.001 0 0 0 0.001 0 0 0 ...
         $ data:List of 4
          ..$ : num [1:256,}
\usage{
CamSizerXT
}
\description{
The dataset comprises a CamSizerXT measurement file.
}
\examples{

data(CamSizerXT)

## plot default data set parmeters (p3 vs. class limits)
plot(CamSizerXT, 
     type = "l", 
     log = "x")

}
\keyword{datasets}
