% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between phi and micrometers.}
\usage{
convert_units(phi, mu)
}
\arguments{
\item{phi}{\code{Numeric} vector, grain-size class values in phi.}

\item{mu}{\code{Numeric} vector, grain-size class values in micrometers.}
}
\value{
\code{Numeric} vector, converted grain-size class values.
}
\description{
The function converts values from the phi-scale to the 
micrometer-scale and vice versa.
}
\examples{

## generate phi-values
phi <- -2:5

## convert and show phi to mu
mu  <- convert_units(phi = phi)
mu

## convert and show phi to mu
convert_units(mu = mu)

}
\author{
Michael Dietze
}
\keyword{grainsize}
