% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize-package.R
\docType{package}
\name{grainsize}
\alias{grainsize}
\alias{grainsize-package}
\title{Grain-size data analysis toolbox}
\description{
Grain-size data analysis functions
}
\author{
Michael Dietze
}
\keyword{grain-size}
