% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsize-package.R
\docType{data}
\name{LA950}
\alias{LA950}
\title{LA950 example data set}
\format{List of 3
         $ x   : num [1:92] 0.02 0.023 0.026 0.03 0.034 0.039 ...
         $ y   : num [1:92] 0 0 0 0 0 0 0 0 0 0 ...
         $ meta:List of 24
          ..$ ID              : chr "S-1"
          ..$ data            : num [1:92] 0 0 0 0 0 0 0 0 0 0 ...
          ..$ unit            : chr "micrometres"}
\usage{
LA950
}
\description{
The dataset comprises a LA950 laser sizer measurement file.
}
\examples{

# read example data set
data(LA950, envir = environment())

# inspect imported data structure
str(LA950)

# plot grain-size distribution curve
plot(LA950, 
     type = "l", 
     log = "x")

}
\keyword{datasets}
