% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CamSizerXT.R
\name{read_CamSizerXT}
\alias{read_CamSizerXT}
\title{Import CamSizerXT measurement files.}
\usage{
read_CamSizerXT(file, output = "p3")
}
\arguments{
\item{file}{\code{Character} value, file name to be imported.}

\item{output}{\code{Character} value, primary output parameter used 
for plotting. One out of \code{p3}, \code{Q3}, \code{xFe3}, 
\code{xMa3}, \code{xc3}, \code{xFe_min3}, \code{xMa_min3}, 
\code{xc_min3}, \code{xFe_max3}, \code{xMa_max3}, \code{xc_max3}. 
The specified parameter must be present in the data set. Only the
data for the first measured data set is used. Default is \code{p3}.}
}
\value{
\code{List} object with imported data. Elements \code{x} and 
\code{y} are grain size class limits and measured content, 
respectively. The object \code{data} contains a list with all 
measured parameters. Object \code{meta} contains meta data.

The CamSizerXT software allows exporting measurement data
as ACII-files. Therefore use Menu > Evaluation > Daily Report, 
select the desired measurement file, toggle all measured 
parameters of interest and use menu File > Export (as *.xle). 
This creates an Excel-readable file, which however cannot be 
imported to R, yet. So it needs to be opened in Excel and 
exported as *.txt-file. In summery, please do not change the 
files. it is the ASCII-file directly generated by Excel that
can be imported to R.
The output is a list object with four elements. \code{x} and 
\code{y} can be directly used to plot the object and will 
contain the primary size definition as set in the input parameters,
by default \code{p3}, and the respective amount of the parameter. The 
element \code{data} is a list object. The CamSizer XT supports more 
than one output file per sample. And thus, several definitions of grain 
size are possible (see \code{x$meta$size.definition}). For each of the 
size definitions as series of grain size and shape parameters are 
measured. The measured parameters are stored in 
\code{x$meta$parameters} and as column names in the data object. Finally,
a meta data object is returned.
}
\description{
This function imports ASCII-output of the Retsch CamSizerXT device.
}
\examples{

## uncomment to use or use example data set just below
## load example data
# x <- read.CamSizerXT(file = "CamSizerXT.txt")

## read example data set
data(CamSizerXT, envir = environment())

## plot default data set parmeters (p3 vs. class limits)
plot(CamSizerXT, type = "l", log = "x")

## show some data
CamSizerXT$meta$ID
CamSizerXT$meta$parameters

## plot grain size distributions of the four size definitions
plot(NA, 
     xlim = range(CamSizerXT$x), 
     ylim = range(CamSizerXT$y), 
     main = "Dune sand, p3-distribution",
     xlab = "Grain size", 
     ylab = "Amount",
     log = "x")

for(i in 1:length(CamSizerXT$meta$size.definition)) {
  lines(x = CamSizerXT$x, 
        y = CamSizerXT$data[[i]][,1], 
        col = i)
}

legend(x = "topleft", 
       legend = CamSizerXT$meta$size.definition, 
       col = 1:4, 
       lty = 1)

## plot grain size distribution and symmetry
plot(CamSizerXT, 
     main = "Dune sand",
     xlab = "Grain size (micrometres)", 
     ylab = "Amount (\%), Symmetry index * 3",
     xlim = c(100, 1000),
     log = "x",
     type = "l")

lines(x = CamSizerXT$x, 
      y = CamSizerXT$data[[1]][,14] * 3, 
      col = "grey")

}
\author{
Michael Dietze
}
\keyword{grainsize}
