% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_LA950.R
\name{read_LA950}
\alias{read_LA950}
\title{Read LA-950 measurement files.}
\usage{
read_LA950(file, version = "8", normalise = TRUE)
}
\arguments{
\item{file}{\code{Character} value, name of the file to be read, 
with extension.}

\item{version}{\code{Character} value, file format version. Default is
\code{"8"}.}

\item{normalise}{\code{Logical} value, option to normalise measured 
class values to the base of 100 percent. Default is \code{TRUE}.}
}
\value{
\code{list} object containing sample ID, grain-size data
and metadata.
}
\description{
This function reads the ASCII output files to R.
}
\examples{

## uncomment to use or use example data set just below
# ## load example data
# x <- read.LA950(file = "LA950.txt")

## read example data set
data(LA950)

# inspect imported data structure
str(LA950)

# plot grain-size distribution curve
plot(LA950, type = "l", log = "x")

# import all files in a folder and write data to a matrix
#   X <- numeric(0)
#   ID <- numeric(0)
#   files <- list.files()
#   for(i in 1:length(files)) {
#     x <- read.LA950(file = files[i])
#     X <- rbind(X, x$y)
#     ID <- c(ID, x$meta$ID)
#   }
#   colnames(X) <- x$x
#   rownames(X) <- ID  

}
\author{
Michael Dietze
}
\keyword{grainsize}
